REFACTOR : 이넘패키지로 분리
