+++ backend/src/main/java/com/example/demo/domain/mail/enums/MailStatus.java
@@ -1,4 +1,4 @@
-package com.example.demo.domain.mail.entity;
+package com.example.demo.domain.mail.enums;
 
 public enum MailStatus {
     PENDING, // 처리 중
